path.h: obstacle_from_disc returns short (path_obstacle_avoidance: 1 function strict exact)
Header packet P5 (150K lane). January's inlined copy in _path_add_steps loads and stores the helper value as 16 bits (mov dx,word [ebx+ecx*8+0xa]; mov word [ebp-0x10],dx) with a dword NONE arm, and the out-of-line _obstacle_from_disc widens an int-typed ?: (movsx eax,word [eax+2] / or eax,-1). The owner prototype and definition now return short, the definition is the single ?: expression, and the redundant (short) cast in path_new is dropped. Declaration count unchanged.

Consumers: 56 transitive path.h includers shadow-compiled H vs P (+1 / 0 losses). Live tree: full ninja + stable sweep +1 gained (_path_add_steps 816, park removed) / 0 regressions; parks 289/0/0.

+++ source/ai/path.h
@@ -432,7 +432,7 @@ void render_debug_obstacles(
 __inline struct obstacle_disc const *obstacles_get_disc(
 	struct obstacles const *obstacles,
 	short disc_index);
-long obstacle_from_disc(
+short obstacle_from_disc(
 	struct obstacles const *obstacles,
 	short disc_index);
 void render_debug_path(
